:art: 优化代码, 完成 `getGJLevelScores211` 和 `getGJLevelScoresPlat` 接口

+++ database/migrations/2024_09_08_145533_create_levels_table_in_endless_server_schema.php
@@ -27,9 +27,9 @@ return new class extends Migration {
 			$table->integer('password')->nullable();
 			$table->foreignIdFor(Level::class, 'original_level_id')->nullable();
 			$table->boolean('two_player_mode_enabled');
-			$table->unsignedSmallInteger('objects');
+			$table->integer('objects');
 			$table->forEnum('coins', GeometryDashLevelCoinCounts::class);
-			$table->unsignedInteger('requested_stars');
+			$table->integer('requested_stars');
 			$table->forEnum('unlisted_type', GeometryDashLevelUnlistedTypes::class);
 			$table->boolean('ldm_enabled');
 			$table->integer('editor_time')->nullable();
